pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "A professional Model Context Protocol (MCP) client with Chainlit UI and Ollama integration."
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
license = { text = "MIT" }
dependencies = [
    "chainlit>=1.3.0",
    "mcp>=1.0.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "tiktoken>=0.7.0",
]

[dependency-groups]
dev = [
    "pytest",
    "black",
    "isort",
    "mypy"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
include = [
    "core/**",
    "interfaces/**",
    "main.py",
    "ui.py",
    "mcp.json",
    "README.md",
    ".env.example"
]

[tool.hatch.build.targets.sdist]
include = [
    "core/**",
    "interfaces/**",
    "main.py",
    "ui.py",
    "mcp.json",
    "README.md",
    ".env.example",
    "pyproject.toml"
]

